% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singstat_endpoint.R
\name{singstat_endpoint}
\alias{singstat_endpoint}
\title{Singstat endpoint URL generator}
\usage{
singstat_endpoint(API_type)
}
\arguments{
\item{API_type}{Compulsory argument. The API endpoint that the URL should be constructed for. Accepts "resourceId", "metadata", or "tabledata".}
}
\value{
The URL of the API endpoint passed as the argument.
}
\description{
Generate the URL of SingStat table builder API endpoints
}
\examples{
\dontrun{
singstat_endpoint("resourceId")
singstat_endpoint("invalid_endpoint") #error}


}
