% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singstat_metadata.R
\name{singstat_metadata}
\alias{singstat_metadata}
\title{Access Singstat Metadata Endpoint}
\usage{
singstat_metadata(
  resourceID,
  print_info = TRUE,
  filter_level = NULL,
  filter_var = NULL,
  ignore_case = TRUE
)
}
\arguments{
\item{resourceID}{Compulsory argument. The resource ID for the table to request metadata for.}

\item{print_info}{Defaults to \code{TRUE}. When \code{TRUE}, prints the metadata for the whole table, also stored as
attributes to the console as a message.}

\item{filter_level}{An integer argument. When not \code{NULL}, filters the dataframe to include only information on
variables at the specified level, or above. See details.}

\item{filter_var}{A regex argument. When not \code{NULL}, filters the dataframe to include only variables with names
containing the regex pattern. See details.}

\item{ignore_case}{Used with filter_var. When \code{TRUE}, filter_var is case-insensitive.}
}
\value{
A dataframe containing metadata on individual variables in a SingStat table, based on its resource ID. In addition to
metadata on individual variables, the metadata API also provides information regarding the table itself. These are
stored as attributes to the returned dataframe. When \code{print_info} is \code{TRUE} (the default), these
attributes are also printed to the console for easy reference as a message.

More details on the attributes containing table information:
\itemize{
 \item{\code{resourceID}} -The resourceID of the table.
 \item{\code{title}} - The table title
 \item{\code{frequency}} - The frequency of measurement
 \item{\code{uom}} - The unit of measure
 \item{\code{datasource}} - The source of the data
 \item{\code{url}} - Direct link to the data on the SingStat Table Builder webclient (not the JSON file)
 \item{\code{footnote}} - Footnotes
 \item{\code{startPeriod}} - The first year of data availability
 \item{\code{endPeriod}} - The most recent year of data availability
 \item{\code{termsOfUse}} - A link to the SingStat Table Builder Terms of Use.
 \item{\code{apiTermsOfService}} - A link to the Singapore Government's API Terms of Service.
}


More details on the information on variables, contained in the dataframe:
\itemize{
 \item{\code{variableCode}} -The code that identifies a specific variable in the table
 \item{\code{variableName}} - The name of the variable
 \item{\code{level}} - Variables in a table are organised in a hierarchicalised manner. This can also be seen
     in the variable codes - child variables have a variable code derived from the parent variable code. The
     levels indicate what level of the hierarchy a variable belongs to, with 1 being the top level variable.
 \item{\code{uom}} - The unit of measure
 \item{\code{footnote}} - Footnotes
}
}
\description{
Constructs a request to the SingStat Table Builder Metadata Endpoint and sends a GET request.
Returns a dataframe containing information about the table variables. Other metadata about the table as a whole
are saved as attributes to this dataframe.
}
\details{
\code{filter_level} accepts an integer argument. If a numeric variable is passed, it is coerced into an integer using
    as.integer().

\code{filter_var} is passed to stringr::str_detect as its pattern argument.

Note that \code{filter_level} is executed before \code{filter_var}.
}
\examples{
singstat_metadata(16112)
singstat_metadata(16112, print_info = FALSE)

singstat_metadata(16112, filter_level = 3)
singstat_metadata(16102, filter_var = "investment")

}
\seealso{
\code{singstat_resource} \code{singstat_tabledata}
}
