% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singstat_resource.R
\name{singstat_resource}
\alias{singstat_resource}
\title{Access Singstat Resource ID Endpoint}
\usage{
singstat_resource(keyword, searchOption = "all", printSummary = FALSE)
}
\arguments{
\item{keyword}{Compulsory argument. The search query.}

\item{searchOption}{Where one wants to search for the keyword. See details.}

\item{printSummary}{Defaults to \code{FALSE}. If set to \code{TRUE}, prints out some details on the search request.
These include the full URL of the request, and some other information provided by the Table Builder.}
}
\value{
A data frame containing the search results. The first column is the resourceID for a SingStat table. The second column
is the table's corresponding title.
}
\description{
Constructs a request to the SingStat Table Builder Resource ID Endpoint and sends a GET request.
Returns a dataframe of the SingStat tables and their corresponding resource IDs that match the search.
}
\details{
If no responses are found for the query, a null object is returned.

For \code{searchOption}, The default option is "all", which searches for the keywords in the both variables
and titles of tables. Other options are "variable", and "title", which searches for the keywords only in variables and titles
respectively.
}
\examples{
singstat_resource("employment")
singstat_resource("gdp", searchOption = "variable", printSummary = TRUE)
\dontrun{
singstat_resource("employment", searchOption = "none") #error}


}
\seealso{
\code{singstat_metadata}, \code{singstat_search}
}
