% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singstat_search.R
\name{singstat_search}
\alias{singstat_search}
\title{Search for SingStat tables}
\usage{
singstat_search(keywords, search_options, search_type = "or")
}
\arguments{
\item{keywords}{A character vector with each search query as an element.}

\item{search_options}{A character vector of the same length as \code{keywords} that provides the corresponding search option
for a \code{singstat_resource} call using that keyword.}

\item{search_type}{A string argument that can be either "and" or "or" (defaults to "or"). If "or", results are returned
if they are found for a search for any of the keywords. If "and", only results that match all valid queries (ones that did not
throw an error) are returned.}
}
\value{
A dataframe containing the aggregated search results. Format differs depending on search type. If "and", return object is similar
    to \code{singstat_resource}. If "or", a new column is made for each keyword showing if the particular table was returned during
    a search for that keyword.
}
\description{
Passes each keyword provided to \code{singstat_resource}, making multiple calls to the resourceID API endpoint
    and aggregating the results
}
\examples{
singstat_search(keywords = c("gdp", "government"),
                search_options = c("all", "all"))

singstat_search(keywords = c("gdp", "government"),
                search_options = c("all", "all"),
                search_type = "and")

}
\seealso{
\code{singstat_resource}
}
