% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singstat_tabledata.R
\name{singstat_tabledata}
\alias{singstat_tabledata}
\title{Access Singstat Tabledata Endpoint}
\usage{
singstat_tabledata(
  resourceID,
  print_info = TRUE,
  variables = NULL,
  between = NULL,
  sort_by = "variableCode asc",
  offset = NULL,
  limit = NULL,
  time_filter = NULL,
  var_search = NULL
)
}
\arguments{
\item{resourceID}{Compulsory argument. The resource ID for the table to request data for.}

\item{print_info}{Defaults to \code{TRUE}. When \code{TRUE}, prints the table metadata for the whole table.}

\item{variables}{A character vector specifying the variable codes one wants to obtain. Defaults to \code{NULL}, in which
case all variables in the table are returned.}

\item{between}{A numeric vector of length 2. Specifies the start and end points (inclusive) of a range of data values
to return. Values outside this range are filtered out Defaults to \code{NULL}, in which no filter is applied.}

\item{sort_by}{The data frame will give each variable, for each year, a single row as per the default object returned by the API.
This will be sorted by variableCode by default in ascending order.
One can also sort by time, value, level, or variable name, in ascending or descending order. See details.}

\item{offset}{Numeric variable. Can specify the first n number of records to be excluded in the returned result. Defaults
to \code{NULL}.}

\item{limit}{the maximum number of records to beincluded in the returned result. Maximum provided by API is 2000.}

\item{time_filter}{A character vector. Returns records of specific time points (not a range) based on the type of selected table.
See details.}

\item{var_search}{A character argument. API will returns only records of variables with a name containing the search string.}
}
\value{
A dataframe containing the requested table with 6 columns. Also provides information regarding the table itself,
stored as attributes to the returned dataframe. When \code{print_info} is \code{TRUE} (the default), these
attributes are also printed to the console for easy reference as a message.

More details on the attributes containing table information:
\itemize{
 \item url - Direct link to the data on the SingStat Table Builder webclient (not the JSON file)
 \item uom - The unit of measure
 \item total - total number of rows
 \item footnote - Footnotes
}
More details on the information on variables, contained in the dataframe:
\itemize{
 \item frequency -The interval of measurement. Monthly, quarterly, half-yearly, or annual
 \item time - the instance of measurement
 \item level - Variables in a table are organised in a hierarchicalised manner. This can also be seen
     in the variable codes - child variables have a variable code derived from the parent variable code. The
     levels indicate what level of the hierarchy a variable belongs to, with 1 being the top level variable.
 \item variableCode -The code that identifies a specific variable in the table
 \item variableName - The name of the variable
 \item value - the value of the measurement
 \item uom - The unit of measure
 \item footnote - Footnotes

}
}
\description{
Constructs a request to the SingStat Table Builder Tabledata Endpoint and sends a GET request.
Returns a dataframe containing the variables and values of the requested table. Selected metadata about the table as
a whole are saved as attributes to this dataframe. This metadata may differ from what is returned by
\code{singstat_metadata} for the same resourceID.
}
\details{
\code{sort_by} accepts a character argument of a specific format. One must specify the column to sort (time, level, value
    value, variableCode, variableName) followed by a space, then asc (for ascending) or desc (for descending). E.g.
    "variableCode asc" or "value desc"

\code{time_filter} accepts a character argument of a specific format. This differs depending on whether the table is monthly,
    half-yearly, quarterly, or annual. Examples below:
    \itemize{
      \item monthly - c("2018 Mar", "2019 Mar")
      \item quarterly - c("2017 4Q", "2018 1Q")
      \item half yearly - c("2017 H1", "2018 H2")
      \item annual - c("2017", "2018")
   }
}
\examples{
singstat_tabledata(16112)
singstat_tabledata(15808, print_info = FALSE, offset = 5)

}
\seealso{
\code{singstat_tidytable}, \code{singstat_metadata}, \code{singstat_craft_table}
}
