% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singstat_tidytable.R
\name{singstat_tidytable}
\alias{singstat_tidytable}
\title{Get a tidy data frame from SingStat}
\usage{
singstat_tidytable(
  resourceID,
  print_info = TRUE,
  ...,
  force_annual = FALSE,
  force_op = "sum",
  na.rm = FALSE
)
}
\arguments{
\item{resourceID}{Compulsory argument. The resource ID for the table to request data for.}

\item{print_info}{Defaults to \code{TRUE}. When \code{TRUE}, prints the table metadata for the whole table.}

\item{...}{Additional arguments to pass to \code{singstat_tabledata}. See documentation for \code{singstat_tabledata} for more
details.}

\item{force_annual}{Defaults to \code{FALSE}. When TRUE for monthly, quarterly, or half-yearly data, sums or finds the mean
values of all the measurements for that year to give an annual value. The frequency column in the table is mutated to "Annual".}

\item{force_op}{String to indicate the aggregation function used in force_annual. Either "sum" or "mean". Defaults to "sum".}

\item{na.rm}{logical. Should missing values (including NaN) be removed? Passed to \code{sum} or \code{mean} for \code{force_annual}.}
}
\value{
A dataframe containing the requested table with time period (year, and div for quarterly, monthly, and half-yearly tables),
    frequency of measurement, and table variables as columns.
}
\description{
Passes arguments to \code{singstat_tabledata}, wrangling its return object into a data frame where the
    time period are observations and the variables are columns.
}
\examples{
singstat_tidytable(16112)
singstat_tidytable(15808, print_info = FALSE)
singstat_tidytable(14506, force_annual = TRUE, na.rm = TRUE)

}
\seealso{
\code{singstat_tabledata}, \code{sum}, \code{mean}
}
